[ADD] Documentação faltante nas headers
